# specs/002-crm-pacientes/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Paciente360 — API v1 (Fase 2 — CRM Pacientes)
  version: 0.2.0
  description: |
    Endpoints novos da Fase 2: gestão de pacientes, convênios, tags,
    anotações, timeline, importação, mesclagem, funil de leads e
    exportação. Todos os endpoints autenticados são escopados ao tenant
    resolvido por subdomínio (herdado da Fase 0).
  contact:
    name: Paciente360 Engineering
servers:
  - url: https://{slug}.crm.com.br/api/v1
    variables:
      slug:
        default: clinica-exemplo
  - url: https://{slug}.lvh.me/api/v1
    variables:
      slug:
        default: clinica-alfa
...
paths:
...
  /pacientes/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [pacientes]
      summary: Ficha completa de paciente
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PacienteResource' }
        '404': { description: Não encontrado ou anonimizado }
    patch:
      tags: [pacientes]
      summary: Atualiza paciente (campos não sensíveis e contatos)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePacienteRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PacienteResource' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [pacientes]
      summary: Soft delete (raramente usado; anonimização é preferida)
      responses:
        '204': { description: Soft-deletado }
        '403': { description: Sem ability paciente.delete }
...
  /pacientes/{id}/timeline:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [timeline]
      summary: Eventos da timeline do paciente em ordem cronológica reversa
      parameters:
        - in: query
          name: tipo
          description: Filtro por tipo de evento (repetível).
          schema: { type: array, items: { type: string } }
        - in: query
          name: cursor
          description: Cursor opaco para paginação cursor-based (mais eficiente que page).
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimelineResponse' }

  # ──────────────────────────────────────────────────────────────
  # Anotações
  # ──────────────────────────────────────────────────────────────
...
  /pacientes/{id}/tags:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      tags: [tags]
      summary: Aplica tag a paciente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_id]
              properties:
                tag_id: { type: integer }
      responses:
        '204': { description: Aplicada (idempotente) }
        '409': { description: 'Limite soft de 10 tags atingido; aplica com warning no header X-Soft-Warning.' }
  /pacientes/{id}/tags/{tag_id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
      - in: path
        name: tag_id
        required: true
        schema: { type: integer }
    delete:
      tags: [tags]
      summary: Remove tag do paciente
      responses:
        '204': { description: Removida }
        '403': { description: 'Tag sistêmica não removível por usuário comum.' }

  # ──────────────────────────────────────────────────────────────
  # Convênios
  # ──────────────────────────────────────────────────────────────
...
components:
  responses:
...
    ValidationError:
      description: Erros de validação (422)
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors:
                type: object
                additionalProperties:
                  type: array
                  items: { type: string }
  schemas:
    PacienteResource:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        cpf: { type: string, nullable: true, description: 'Formato 000.000.000-00 quando preenchido' }
        documento_estrangeiro: { type: string, nullable: true }
        data_nascimento: { type: string, format: date, nullable: true }
        idade: { type: integer, nullable: true, description: 'Calculada server-side a partir de data_nascimento' }
        telefone_primario: { type: string, nullable: true, description: 'Formato (00) 00000-0000' }
        telefones_secundarios:
          type: array
          items:
            type: object
            properties:
              numero: { type: string }
              label: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        endereco:
          type: object
          nullable: true
          properties:
            cep: { type: string }
            logradouro: { type: string }
            numero: { type: string }
            complemento: { type: string, nullable: true }
            bairro: { type: string }
            cidade: { type: string }
            uf: { type: string, minLength: 2, maxLength: 2 }
        status: { type: string, enum: [lead, ativo, inativo, bloqueado] }
        origem: { type: string, enum: [site, indicacao, whatsapp, instagram, telefone, presencial, outro] }
        origem_detalhe: { type: string, nullable: true }
        origem_origem: { type: string, enum: [manual, canal] }
        profissional_responsavel:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            nome: { type: string }
        tags:
          type: array
          items: { $ref: '#/components/schemas/TagResource' }
        convenios:
          type: array
          items: { $ref: '#/components/schemas/PacienteConvenioResource' }
        funil:
          type: object
          nullable: true
          properties:
            coluna: { $ref: '#/components/schemas/FunilColunaResource' }
            posicao: { type: number, format: double }
        anonimizado_em: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CreatePacienteRequest:
      type: object
      required: [nome]
      properties:
        nome: { type: string, minLength: 2, maxLength: 150 }
        cpf: { type: string, description: 'Opcional. 11 dígitos com ou sem máscara; validado por DV server-side.' }
        documento_estrangeiro: { type: string, maxLength: 30 }
        data_nascimento: { type: string, format: date }
        telefone_primario: { type: string }
        telefones_secundarios:
          type: array
          items:
            type: object
            properties:
              numero: { type: string }
              label: { type: string }
        email: { type: string, format: email }
        endereco:
          type: object
          properties:
            cep: { type: string }
            logradouro: { type: string }
            numero: { type: string }
            complemento: { type: string }
            bairro: { type: string }
            cidade: { type: string }
            uf: { type: string }
        status:
          type: string
          enum: [lead, ativo]
          default: lead
          description: 'Apenas lead ou ativo no momento da criação.'
        origem:
          type: string
          enum: [site, indicacao, whatsapp, instagram, telefone, presencial, outro]
          default: outro
        origem_detalhe: { type: string, maxLength: 255 }
        profissional_responsavel_id: { type: integer, nullable: true }
        convenios:
          type: array
          maxItems: 2
          items:
            type: object
            required: [convenio_id, papel]
            properties:
              convenio_id: { type: integer }
              papel: { type: string, enum: [principal, secundario] }
              numero_carteirinha: { type: string, maxLength: 30 }
        tags:
          type: array
          items: { type: integer, description: 'tag_id existente; criar nova via POST /tags antes.' }
        ignorar_duplicata: { type: boolean, default: false, description: 'Se true, cria mesmo havendo duplicata por CPF.' }
    UpdatePacienteRequest:
      allOf:
        - $ref: '#/components/schemas/CreatePacienteRequest'
        - type: object
          description: 'Mesmos campos do create, todos opcionais. CPF não é alterável após criação (apenas via mesclagem ou anonimização).'
...
    TagResource:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        nome_normalizado: { type: string }
        tipo: { type: string, enum: [livre, sistemica] }
        cor: { type: string, nullable: true }
        descricao: { type: string, nullable: true }
        pacientes_count: { type: integer, description: 'Contagem de pacientes com a tag. Pode ser omitido em endpoints list para performance.' }
    ConvenioResource:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        codigo_ans: { type: string, nullable: true }
        is_active: { type: boolean }
        pacientes_count: { type: integer }
        created_at: { type: string, format: date-time }
...
    PacienteConvenioResource:
      type: object
      properties:
        id: { type: integer }
        convenio: { $ref: '#/components/schemas/ConvenioResource' }
        numero_carteirinha: { type: string, nullable: true }
        papel: { type: string, enum: [principal, secundario] }
    FunilColunaResource:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        slug: { type: string }
        posicao: { type: integer }
        cor: { type: string, nullable: true }
        is_terminal: { type: boolean }
        motivo_obrigatorio: { type: boolean }
        is_system: { type: boolean }
        pacientes_count: { type: integer }
    TimelineResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/EventoTimelineResource' }
        meta:
          type: object
          properties:
            next_cursor: { type: string, nullable: true }
            has_more: { type: boolean }
    EventoTimelineResource:
      type: object
      properties:
        id: { type: integer }
        tipo: { type: string }
        actor:
          type: object
          properties:
            type: { type: string, enum: [user, system, webhook, ia] }
            user_id: { type: integer, nullable: true }
            user_name: { type: string, nullable: true }
        payload: { type: object, additionalProperties: true }
        referencia:
          type: object
          nullable: true
          properties:
            tipo: { type: string }
            id: { type: integer }
        created_at: { type: string, format: date-time }
...